chat input style optimizations
