+++ src/views/components/chat-input/tools/emoji.tsx
@@ -2,10 +2,12 @@ import React from 'react';
 import Box from '@mui/material/Box';
 import EmojiPicker from 'components/emoji-picker';
 import usePopover from 'hooks/use-popover';
+import useToolStyle from 'views/components/chat-input/tools/style';
 import EmoticonHappy from 'svg/emoticon-happy';
 
 const Emoji = (props: { onSelect: (selected: string) => void }) => {
     const [, showPopover] = usePopover();
+    const toolSx = useToolStyle();
 
     const emojiClicked = (event: React.MouseEvent<HTMLDivElement>) => {
         showPopover({
@@ -19,10 +21,7 @@ const Emoji = (props: { onSelect: (selected: string) => void }) => {
         });
     }
 
-    return <Box sx={{
-        display: 'inline-flex',
-        alignItems: 'center'
-    }} onClick={emojiClicked}>
+    return <Box sx={toolSx} onClick={emojiClicked}>
         <EmoticonHappy height={20}/>
     </Box>;
 }
