+++ src/views/components/chat-input/tools/index.tsx
@@ -1,13 +1,10 @@
 import {MutableRefObject, useEffect} from 'react';
 import Emoji from 'views/components/chat-input/tools/emoji';
 import Gif from 'views/components/chat-input/tools/gif';
-import Box from '@mui/material/Box';
-import {useTheme} from '@mui/material/styles';
 import {focusElem, insertText} from 'util/dom';
 
 const Tools = (props: { inputRef: MutableRefObject<HTMLDivElement | null>, senderFn: (message: string) => void }) => {
     const {inputRef} = props;
-    const theme = useTheme();
 
     useEffect(() => {
         const onPaste = (e: ClipboardEvent) => {
@@ -29,31 +26,14 @@ const Tools = (props: { inputRef: MutableRefObject<HTMLDivElement | null>, sende
         }
     }, []);
 
-    const toolSx = {
-        display: 'flex',
-        alignItems: 'center',
-        padding: '6px',
-        cursor: 'pointer',
-        mr: '6px',
-        borderRadius: theme.shape.borderRadius,
-        ':hover': {
-            background: theme.palette.divider,
-        }
-    };
-
     return <>
-        <Box sx={toolSx}>
-            <Emoji onSelect={(emoji) => {
-                focusElem(inputRef.current!);
-                insertText(emoji);
-            }}/>
-        </Box>
-
-        <Box sx={toolSx}>
-            <Gif onSelect={(gif) => {
-                props.senderFn(gif);
-            }}/>
-        </Box>
+        <Emoji onSelect={(emoji) => {
+            focusElem(inputRef.current!);
+            insertText(emoji);
+        }}/>
+        <Gif onSelect={(gif) => {
+            props.senderFn(gif);
+        }}/>
     </>;
 }
